+++ packages/e2e/cypress/integration/explore.spec.ts
@@ -1,15 +1,16 @@
 describe('Explore', () => {
+    before(() => {
+        // @ts-ignore
+        cy.login();
+    });
+
+    beforeEach(() => {
+        Cypress.Cookies.preserveOnce('connect.sid');
+    });
+
     it('Should query orders', () => {
         cy.visit('/');
 
-        // Close query dev tools
-        cy.get('body').then((body) => {
-            if (body.find('.ReactQueryDevtools').length > 0) {
-                cy.get('.ReactQueryDevtools')
-                    .findByRole('button', { name: /Close/i })
-                    .click();
-            }
-        });
         cy.findByText('Orders').click();
         cy.findByText('First name').click();
         cy.findByText('Total orders').click();